pyproject.toml
[project]
name = "jira-mcp"
version = "0.1.0"
description = "MCP server for Jira Cloud reporting via OAuth 2.0"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "truststore>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
jira-mcp = "jira_mcp.server:main"
jira-mcp-login = "jira_mcp.auth:login"
